remove move selector config
