+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/config/ruin/RuinPhaseConfig.java
@@ -5,11 +5,9 @@ import java.util.function.Consumer;
 
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlElementWrapper;
-import javax.xml.bind.annotation.XmlTransient;
 import javax.xml.bind.annotation.XmlType;
 
 import org.optaplanner.core.config.heuristic.selector.entity.EntitySelectorConfig;
-import org.optaplanner.core.config.heuristic.selector.move.generic.RuinMoveSelectorConfig;
 import org.optaplanner.core.config.phase.PhaseConfig;
 import org.optaplanner.core.config.solver.termination.TerminationConfig;
 import org.optaplanner.core.config.util.ConfigUtils;
@@ -23,9 +21,6 @@ import org.optaplanner.core.config.util.ConfigUtils;
 public class RuinPhaseConfig extends PhaseConfig<RuinPhaseConfig> {
 
     public static final String XML_ELEMENT_NAME = "ruin";
-
-    @XmlTransient
-    private RuinMoveSelectorConfig moveSelectorConfig = null;
 
     @XmlElement(name = "entitySelector")
     private EntitySelectorConfig entitySelectorConfig = null;
@@ -54,6 +49,22 @@ public class RuinPhaseConfig extends PhaseConfig<RuinPhaseConfig> {
         this.variableNameIncludeList = variableNameIncludeList;
     }
 
+    public Integer getPercentageToRuin() {
+        return percentageToRuin;
+    }
+
+    public EntitySelectorConfig getEntitySelectorConfig() {
+        return entitySelectorConfig;
+    }
+
+    public EntitySelectorConfig getSecondaryEntitySelectorConfig() {
+        return secondaryEntitySelectorConfig;
+    }
+
+    public List<String> getVariableNameIncludeList() {
+        return variableNameIncludeList;
+    }
+
     public void setPercentageToRuin(Integer percentageToRuin) {
         this.percentageToRuin = percentageToRuin;
     }
@@ -61,14 +72,6 @@ public class RuinPhaseConfig extends PhaseConfig<RuinPhaseConfig> {
     // ************************************************************************
     // Constructors and simple getters/setters
     // ************************************************************************
-
-    public RuinMoveSelectorConfig getMoveSelectorConfig() {
-        if (moveSelectorConfig == null) {
-            moveSelectorConfig = new RuinMoveSelectorConfig(entitySelectorConfig, secondaryEntitySelectorConfig,
-                    variableNameIncludeList, percentageToRuin);
-        }
-        return moveSelectorConfig;
-    }
 
     @Override
     public TerminationConfig getTerminationConfig() {
@@ -113,8 +116,6 @@ public class RuinPhaseConfig extends PhaseConfig<RuinPhaseConfig> {
         variableNameIncludeList = ConfigUtils.inheritMergeableListProperty(
                 variableNameIncludeList, inheritedConfig.variableNameIncludeList);
         percentageToRuin = ConfigUtils.inheritOverwritableProperty(percentageToRuin, inheritedConfig.percentageToRuin);
-        moveSelectorConfig = ConfigUtils.inheritOverwritableProperty(
-                getMoveSelectorConfig(), inheritedConfig.moveSelectorConfig);
         return this;
     }
 
@@ -125,8 +126,11 @@ public class RuinPhaseConfig extends PhaseConfig<RuinPhaseConfig> {
 
     @Override
     public void visitReferencedClasses(Consumer<Class<?>> classVisitor) {
-        if (moveSelectorConfig != null) {
-            moveSelectorConfig.visitReferencedClasses(classVisitor);
+        if (entitySelectorConfig != null) {
+            entitySelectorConfig.visitReferencedClasses(classVisitor);
+        }
+        if (secondaryEntitySelectorConfig != null) {
+            secondaryEntitySelectorConfig.visitReferencedClasses(classVisitor);
         }
     }
 }
