+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/heuristic/selector/move/MoveSelectorFactory.java
@@ -10,7 +10,6 @@ import org.optaplanner.core.config.heuristic.selector.move.factory.MoveListFacto
 import org.optaplanner.core.config.heuristic.selector.move.generic.ChangeMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.PillarChangeMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.PillarSwapMoveSelectorConfig;
-import org.optaplanner.core.config.heuristic.selector.move.generic.RuinMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.SwapMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.chained.KOptMoveSelectorConfig;
 import org.optaplanner.core.config.heuristic.selector.move.generic.chained.SubChainChangeMoveSelectorConfig;
@@ -29,7 +28,6 @@ import org.optaplanner.core.impl.heuristic.selector.move.factory.MoveListFactory
 import org.optaplanner.core.impl.heuristic.selector.move.generic.ChangeMoveSelectorFactory;
 import org.optaplanner.core.impl.heuristic.selector.move.generic.PillarChangeMoveSelectorFactory;
 import org.optaplanner.core.impl.heuristic.selector.move.generic.PillarSwapMoveSelectorFactory;
-import org.optaplanner.core.impl.heuristic.selector.move.generic.RuinMoveSelectorFactory;
 import org.optaplanner.core.impl.heuristic.selector.move.generic.SwapMoveSelectorFactory;
 import org.optaplanner.core.impl.heuristic.selector.move.generic.chained.KOptMoveSelectorFactory;
 import org.optaplanner.core.impl.heuristic.selector.move.generic.chained.SubChainChangeMoveSelectorFactory;
@@ -78,8 +76,6 @@ public interface MoveSelectorFactory<Solution_> {
             return new KOptMoveSelectorFactory<>((KOptMoveSelectorConfig) moveSelectorConfig);
         } else if (KOptListMoveSelectorConfig.class.isAssignableFrom(moveSelectorConfig.getClass())) {
             return new KOptListMoveSelectorFactory<>((KOptListMoveSelectorConfig) moveSelectorConfig);
-        } else if (RuinMoveSelectorConfig.class.isAssignableFrom(moveSelectorConfig.getClass())) {
-            return new RuinMoveSelectorFactory<>((RuinMoveSelectorConfig) moveSelectorConfig);
         } else {
             throw new IllegalArgumentException(String.format("Unknown %s type: (%s).",
                     MoveSelectorConfig.class.getSimpleName(), moveSelectorConfig.getClass().getName()));
