+++ core/optaplanner-core-impl/src/main/java/org/optaplanner/core/impl/heuristic/selector/move/generic/RuinMoveSelectorFactory.java
@@ -7,25 +7,26 @@ import java.util.Objects;
 import org.optaplanner.core.config.heuristic.selector.common.SelectionCacheType;
 import org.optaplanner.core.config.heuristic.selector.common.SelectionOrder;
 import org.optaplanner.core.config.heuristic.selector.entity.EntitySelectorConfig;
-import org.optaplanner.core.config.heuristic.selector.move.generic.RuinMoveSelectorConfig;
+import org.optaplanner.core.config.ruin.RuinPhaseConfig;
 import org.optaplanner.core.impl.domain.entity.descriptor.EntityDescriptor;
 import org.optaplanner.core.impl.domain.variable.descriptor.GenuineVariableDescriptor;
 import org.optaplanner.core.impl.domain.variable.descriptor.ShadowVariableDescriptor;
 import org.optaplanner.core.impl.heuristic.HeuristicConfigPolicy;
 import org.optaplanner.core.impl.heuristic.selector.entity.EntitySelector;
 import org.optaplanner.core.impl.heuristic.selector.entity.EntitySelectorFactory;
-import org.optaplanner.core.impl.heuristic.selector.move.AbstractMoveSelectorFactory;
 import org.optaplanner.core.impl.heuristic.selector.move.MoveSelector;
+import org.optaplanner.core.impl.util.DescriptorsDeducer;
 
-public class RuinMoveSelectorFactory<Solution_>
-        extends AbstractMoveSelectorFactory<Solution_, RuinMoveSelectorConfig> {
+public class RuinMoveSelectorFactory<Solution_> {
+    private final RuinPhaseConfig config;
+    private final DescriptorsDeducer<Solution_, RuinPhaseConfig> descriptorsDeducer;
 
-    public RuinMoveSelectorFactory(RuinMoveSelectorConfig moveSelectorConfig) {
-        super(moveSelectorConfig);
+    public RuinMoveSelectorFactory(RuinPhaseConfig ruinPhaseConfig) {
+        this.config = ruinPhaseConfig;
+        this.descriptorsDeducer = new DescriptorsDeducer<>(ruinPhaseConfig);
     }
 
-    @Override
-    protected MoveSelector<Solution_> buildBaseMoveSelector(HeuristicConfigPolicy<Solution_> configPolicy,
+    public MoveSelector<Solution_> buildRuinMoveSelector(HeuristicConfigPolicy<Solution_> configPolicy,
             SelectionCacheType minimumCacheType, boolean randomSelection) {
         EntitySelectorConfig entitySelectorConfig =
                 Objects.requireNonNullElseGet(config.getEntitySelectorConfig(), EntitySelectorConfig::new);
@@ -42,10 +43,10 @@ public class RuinMoveSelectorFactory<Solution_>
         EntityDescriptor<Solution_> entityDescriptor = leftEntitySelector.getEntityDescriptor();
 
         List<GenuineVariableDescriptor<Solution_>> variableDescriptorList =
-                deduceVariableDescriptorList(entityDescriptor, config.getVariableNameIncludeList());
+                descriptorsDeducer.deduceVariableDescriptorList(entityDescriptor, config.getVariableNameIncludeList());
 
         Collection<ShadowVariableDescriptor<Solution_>> shadowVariableDescriptorList =
-                deduceShadowVariableDescriptorList(entityDescriptor, config.getVariableNameIncludeList());
+                descriptorsDeducer.deduceShadowVariableDescriptorList(entityDescriptor, config.getVariableNameIncludeList());
         return new CompositeRuinMoveSelector<>(leftEntitySelector, rightEntitySelector, variableDescriptorList,
                 shadowVariableDescriptorList, config.getPercentageToRuin());
     }
