--- core/optaplanner-core-impl/src/test/java/org/optaplanner/core/impl/heuristic/selector/move/generic/RuinMoveSelectorFactoryTest.java
@@ -1,29 +1,0 @@
-package org.optaplanner.core.impl.heuristic.selector.move.generic;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.optaplanner.core.impl.heuristic.HeuristicConfigPolicyTestUtils.buildHeuristicConfigPolicy;
-
-import org.junit.jupiter.api.Test;
-import org.optaplanner.core.config.heuristic.selector.common.SelectionCacheType;
-import org.optaplanner.core.config.heuristic.selector.common.SelectionOrder;
-import org.optaplanner.core.config.heuristic.selector.move.generic.RuinMoveSelectorConfig;
-import org.optaplanner.core.impl.domain.solution.descriptor.SolutionDescriptor;
-import org.optaplanner.core.impl.heuristic.selector.move.MoveSelector;
-import org.optaplanner.core.impl.heuristic.selector.move.MoveSelectorFactory;
-import org.optaplanner.core.impl.testdata.domain.multivar.TestdataMultiVarSolution;
-
-class RuinMoveSelectorFactoryTest {
-
-    @Test
-    void checkDefaultMoveSelector() {
-        SolutionDescriptor solutionDescriptor = TestdataMultiVarSolution.buildSolutionDescriptor();
-        RuinMoveSelectorConfig moveSelectorConfig = new RuinMoveSelectorConfig(null, null, null, 10);
-        MoveSelector moveSelector =
-                MoveSelectorFactory.create(moveSelectorConfig).buildMoveSelector(
-                        buildHeuristicConfigPolicy(solutionDescriptor), SelectionCacheType.JUST_IN_TIME, SelectionOrder.RANDOM);
-        assertThat(moveSelector)
-                .isInstanceOf(CompositeRuinMoveSelector.class);
-        assertTrue(moveSelector.isCountable());
-    }
-}
